Create the files for the project
Create the index.html and js.js files. Insert the standart HTML code
into the index.html file and link the js.js to the index.html using
<script src=''></script> tags.
Create the variable 'choice' and assign it to the starting value of 0
Implement the function getComputerChoice that gets the random
number, then multiplies the number by 3 so that we get the random numbers
in the range of 0-2, and then rounds it with the Math.floor() function.
After that it assigns the result to the 'choice' variable.
Create the if.. else statements in order for the program to return the
values 'Rock', 'Paper', 'Scissors' if the 'choice' is equal to their
value of 0 for 'Rock', 1 for 'Paper' and 2 for 'Scissors';