Implement the code that plays one round of a game
Create the variable 'computerSelection' and set it to the result of calling
'getComputerSelection()' function.
Create the a variable 'computerSelectionLoweCase' to be the lowercased
version of the computerSelection, in order for the code to check all
possible variations of cases in the computerSelection string.
Create the variable 'playerSelection' and set it to be the lowercased
version of the string 'rock'
Create a function 'playARound()' that plays one round of the game.
Inside of the 'playARound()' function, create another function called
'tieLoseWinIndex()' in order to improve the readability of the code.
The function returns an integer from -1 to 1 that represents ; a loss
for the player, a tie, or a victory for the player.
Create the variable 'index' and set to be equal the result of the calling
the function 'tieLoseWinIndex()' with arguments 'computerSelectionLoweCase'
and 'playerSelection'.
Create a check for the value of the index to return the end message.